Fixing library compiling and linking.
Updating all CMakeLists.txt to not include the project directive, realized
that that should only be used once per project and since everything here
is in one project it should only be in one CMakeLists.txt.

Adding CMakeLists.txt to all file directories. They are not all implemented
fully but they are all being called in the hierarchy even if blank. I'm not
sure how this work out in practice for things like papers but I put it there
just in case it could help out. I think I might keep the papers dirs using a
handwritten makefile for now because it is just easier that way and involves
less steps.

Updating source files to make called the GarthLib, they work! I did not have
the headers linked correctly so it was failing before. Libraries are good to
go now.