Creating new python build script.
Created new python build script so that we can move towards being more
cross-platform. I have tried to create it so that, up until now, it should
work just fine on Windows or Unix (including MAC).

Updating .travis.yml file to call python build script instead of shell
script. I might gave to edit it some to install python3 if the server
does not already have it. I don't know of a good way to check which
version is running on the servers. It is probably on some website
somewhere but who has time for that? Committing and letting CI tell me if
I need to update. Shitty but whatever.