Cleaing up .travis.yml
Took me a little while to get this new colorama python module to work
on the Travis servers. I had to do some nasty stuff to get it right but
it works so I'm happy. I don't know much about using the virtualenvs for
different python versions/modules so I'm going to have to look into that
later. The main problem I am having is that pip installs everything under
python2 and not python3. It seems to be a problem for many people mainly
because python3 is still not totally supported with all the setup tools
like pip. Workaround is to just install the modules under 2.7 and manually
copy then to 3.X (3.2 on Travis, 3.3 on my machine).