Cleaning up build script.
I was doing some pretty goofy stuff that should have been abstracted out
into their own functions (like performing system calls). The code is a lot
cleaner now with repect to building on each platform.

I am creating a 'completeScript' function as well that should always be called
when you want to exit, it takes the exit code as an arg and allows for uniformly
printing out an exit message.