Adding in MPI support and normalizing GarthLib.
Not sure if this is a great idea or not but I have normalized the classes
into their own files. Compile time does take a hit now (a few seconds) so
I might undo this change eventually. However, everything compiles fine so
I'm leaving it.

Updating CMakeFiles.txt to check for MPI.

Updating app to use new lib structure.