Trying to get CMake+MPI to work... Shit is rough.
MPI+cmake is really pissing me off! Most people use the mpiXX compiler
wrappers but with cmake it should take care of all the depends correctly.
But... it isn't doing that on all systems. Seems to be working half way
on my machine but not on the Travis machines. We'll see if this works...