Fixing build script.
I had incorrectly updated the build script when messing with
getting the number of processers so that I could pass that to make. I
was originally giving the command an embedded shell command to run which
meant I had to set shell=True for subprocess.call to work correctly.
However, I abstracted that away and calculated that in another place but
never removed the shell=True. This was causing really weird results like
make doc, test, install did not run and the exes were being linked 3
time...? I have no idea, but the issue is resolved now by removing
shell=true and make runs just fine with the -j option.