Adding in -j option to the build script.
Before this option was hardcoded as always on and just looked at the
machine to figure out how many processors are available. I have now set
this as a configurable option on the command line of the build script.
If nothing is supplied, the number will be gotten from the machine. If
-j <int> is supplied, that number will be used. 0 is the default value
and triggers the searching of the machine.